refactor: add sample for navigation
